feat: functional to oop & minor changes;
