+++ dags/hivemind_etl_helpers/github_etl.py
@@ -3,11 +3,11 @@ from datetime import datetime
 
 from dotenv import load_dotenv
 from hivemind_etl_helpers.src.db.github.extract import (
-    fetch_comments,
     fetch_commits,
     fetch_issues,
     fetch_pull_requests,
 )
+from hivemind_etl_helpers.src.db.github.extract import GithubExtraction
 from hivemind_etl_helpers.src.db.github.github_organization_repos import (
     get_github_organization_repos,
 )
@@ -18,7 +18,6 @@ from hivemind_etl_helpers.src.db.github.load import (
 from hivemind_etl_helpers.src.db.github.transform import GitHubTransformation
 from llama_index.core import Document
 from tc_hivemind_backend.db.pg_db_utils import setup_db
-
 
 def process_github_vectorstore(
     community_id: str, github_org_id: str, from_starting_date: datetime | None = None
@@ -61,7 +60,7 @@ def process_github_vectorstore(
     logging.info(f"{len(repository_ids)} repositories to fetch data from!")
 
     # EXTRACT
-    github_comments = fetch_comments(repository_id=repository_ids, from_date=from_date)
+    github_comments = GithubExtraction().fetch_comments(repository_id=repository_ids, from_date=from_date)
     github_commits = fetch_commits(repository_id=repository_ids, from_date=from_date)
     github_issues = fetch_issues(repository_id=repository_ids, from_date=from_date)
     github_prs = fetch_pull_requests(
