+++ dags/hivemind_etl_helpers/src/db/github/extract/__init__.py
@@ -1,5 +1,6 @@
 # flake8: noqa
-from .comments import fetch_comments
 from .commit import fetch_commits
 from .issues import fetch_issues
 from .pull_requests import fetch_pull_requests
+from .github_comment_extraction import GitHubCommentExtraction
+from .github_extraction import GithubExtraction
