+++ dags/hivemind_etl_helpers/tests/integration/test_github_etl_fetch_comments.py
@@ -2,8 +2,7 @@ from datetime import datetime
 from unittest import TestCase
 
 from github.neo4j_storage.neo4j_connection import Neo4jConnection
-from hivemind_etl_helpers.src.db.github.extract import fetch_comments
-
+from hivemind_etl_helpers.src.db.github.extract import GithubExtraction
 
 class TestGithubETLFetchComments(TestCase):
     def setUp(self) -> None:
@@ -11,17 +10,16 @@ class TestGithubETLFetchComments(TestCase):
         self.neo4j_driver = neo4j_connection.connect_neo4j()
         with self.neo4j_driver.session() as session:
             session.execute_write(lambda tx: tx.run("MATCH (n) DETACH DELETE (n)"))
+        self.extractor = GithubExtraction()
 
     def test_get_empty_results_no_from_date(self):
         repository_ids = [123, 124]
-        comments = fetch_comments(repository_id=repository_ids, from_date=None)
+        comments = self.extractor.fetch_comments(repository_id=repository_ids, from_date=None)
         self.assertEqual(comments, [])
 
     def test_get_empty_results(self):
         repository_ids = [123, 124]
-        comments = fetch_comments(
-            repository_id=repository_ids, from_date=datetime(2024, 1, 1)
-        )
+        comments = self.extractor.fetch_comments(repository_id=repository_ids, from_date=datetime(2024, 1, 1))
         self.assertEqual(comments, [])
 
     def test_get_single_comment_single_repo_no_from_date(self):
@@ -56,9 +54,7 @@ class TestGithubETLFetchComments(TestCase):
             )
 
         repository_ids = [123]
-        comments = fetch_comments(
-            repository_id=repository_ids,
-        )
+        comments = self.extractor.fetch_comments(repository_id=repository_ids)
 
         self.assertEqual(len(comments), 1)
         self.assertEqual(comments[0].id, 111)
@@ -137,10 +133,7 @@ class TestGithubETLFetchComments(TestCase):
             )
 
         repository_ids = [123]
-        comments = fetch_comments(
-            repository_id=repository_ids,
-            from_date=datetime(2024, 1, 1),
-        )
+        comments = self.extractor.fetch_comments(repository_id=repository_ids, from_date=datetime(2024, 1, 1))
 
         self.assertEqual(len(comments), 1)
         self.assertEqual(comments[0].id, 111)
