+++ dags/hivemind_github_etl.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import logging
 from datetime import datetime
-
 from airflow import DAG
 from airflow.decorators import task
 from hivemind_etl_helpers.github_etl import process_github_vectorstore
