+++ docker-compose.yaml
@@ -119,6 +119,9 @@ services:
 
   neo4j:
     image: "neo4j:5.9.0"
+    ports:
+      - "7474:7474" # HTTP
+      - "7687:7687" # Bolt
     environment:
       - NEO4J_AUTH=neo4j/neo4j123456
       - NEO4J_PLUGINS=["apoc", "graph-data-science"]
